detail screen with ingredients

+++ remote/src/main/java/com/shapeide/rasadesa/remote/domain/IngredientsModel.kt
@@ -6,10 +6,10 @@ import com.google.gson.annotations.SerializedName
 data class IngredientsModel (
 
   @SerializedName("text"         ) var text         : String? = null,
-  @SerializedName("quantity"     ) var quantity     : Int?    = null,
+  @SerializedName("quantity"     ) var quantity     : Float?    = null,
   @SerializedName("measure"      ) var measure      : String? = null,
   @SerializedName("food"         ) var food         : String? = null,
-  @SerializedName("weight"       ) var weight       : Int?    = null,
+  @SerializedName("weight"       ) var weight       : Double?    = null,
   @SerializedName("foodCategory" ) var foodCategory : String? = null,
   @SerializedName("foodId"       ) var foodId       : String? = null,
   @SerializedName("image"        ) var image        : String? = null
